requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==22.2.0
cached-property==1.5.2
cachetools==5.3.0
charset-normalizer==3.0.1
chex==0.1.6
clu==0.0.6
colorama==0.4.6
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
dm-tree==0.1.8
einops==0.6.1
etils==1.0.0
flatbuffers==1.12
flax==0.6.5
fonttools==4.38.0
future==0.18.3
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
jax==0.4.4
jaxlib==0.4.4+cuda11.cudnn82
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
ml-collections==0.1.0
ml-dtypes==0.0.4
msgpack==1.0.4
numpy==1.22.0
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.3
orbax==0.1.2
packaging==23.0
pandas==1.5.0
Pillow==9.4.0
promise==2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-addons==0.19.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
tensorstore==0.1.32
termcolor==2.2.0
toolz==0.12.0
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
